apps/tracklab-backend/pyproject.toml
[tool.coverage.run]
branch = true
source = [ "src" ]

[tool.coverage.report]
exclude_lines = ['if TYPE_CHECKING:']
show_missing = true

[tool.pytest.ini_options]
addopts = """
 --cov
 --cov-report html:'../../coverage/apps/tracklab-backend/html'
 --cov-report xml:'../../coverage/apps/tracklab-backend/coverage.xml'
 --html='../../reports/apps/tracklab-backend/unittests/html/index.html'
 --junitxml='../../reports/apps/tracklab-backend/unittests/junit.xml'
"""

[project]
name = "tracklab"
version = "1.0.0"
description = "Automatically generated by Nx."
requires-python = ">=3.9,<4"
readme = 'README.md'
dependencies = [
  "fastf1>=3.5.3",
  "grpcio>=1.71.0",
  "grpcio-tools>=1.71.0",
  "protoletariat>=3.3.10",
  "python-dotenv>=1.1.0",
]


[build-system]
requires = ["uv_build>=0.8.9,<0.9.0"]
build-backend = "uv_build"

[tool.ruff]
exclude = [
  ".ruff_cache",
  ".svn",
  ".tox",
  ".venv",
  "dist",
]

line-length = 88
indent-width = 4

[tool.ruff.lint]
select = [
  # pycodestyle
  "E",
  # Pyflakes
  "F",
  # pyupgrade
  "UP",
  # flake8-bugbear
  "B",
  # flake8-simplify
  "SIM",
  # isort
  "I",
]
ignore = []

fixable = ["ALL"]
unfixable = []
